requirements.txt
aiogram==3.2.0
redis==5.0.1
python-dotenv==1.0.1